requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
easyocr==1.7.2
eval_type_backport==0.2.2
fastapi==0.115.12
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.3.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.1
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
openai==1.91.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.0
pillow==11.1.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyparsing==3.2.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.2
selenium==4.31.0
setuptools==76.1.0
shapely==2.0.7
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.2
stringzilla==3.12.3
sympy==1.13.1
tabulate==0.9.0
tifffile==2025.3.13
together==1.5.17
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.4
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
pymongo==4.6.1
pydantic-settings==2.1.0
pyzbar[scripts]